pyproject.toml
[project]
name = "ceu-ai-engineering"
version = "1.0.0"
requires-python = ">=3.13,<3.14"
dependencies = [
    "openai-agents==0.10.5",
    "litellm==1.81.14",
    "boto3==1.42.54",
    "jupyter==1.1.1",
    "chromadb==1.0.21",
    "pysqlite3==0.6.0",
    "chainlit==2.10.0",
    "mcp[cli]==1.26.0",
    "absl-py==2.4.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anthropic==0.84.0",
    "anyio==4.12.1",
    "asyncer==0.0.17",
    "attrs==25.4.0",
    "audioop-lts==0.2.2",
    "bidict==0.23.1",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.5",
    "chevron==0.14.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cryptography==46.0.5",
    "cuda-bindings==12.9.4",
    "cuda-pathfinder==1.4.1",
    "cuid==0.4",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.3.1",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "fastapi==0.135.1",
    "filelock==3.25.0",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "fonttools==4.61.1",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "googleapis-common-protos==1.73.0",
    "griffe==1.15.0",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "hf-xet==1.3.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.6.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "inflection==0.5.1",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kiwisolver==1.4.9",
    "lazify==0.4.0",
    "literalai==0.1.201",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "matplotlib==3.10.8",
    "mdurl==0.1.2",
    "mediapipe==0.10.32",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "numpy==2.4.2",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.5",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvshmem-cu12==3.4.5",
    "nvidia-nvtx-cu12==12.8.90",
    "openai==2.26.0",
    "opencv-contrib-python==4.13.0.92",
    "opencv-python==4.13.0.92",
    "opencv-python-headless==4.13.0.92",
    "opentelemetry-api==1.40.0",
    "opentelemetry-exporter-otlp-proto-common==1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.40.0",
    "opentelemetry-exporter-otlp-proto-http==1.40.0",
    "opentelemetry-instrumentation==0.61b0",
    "opentelemetry-instrumentation-agno==0.53.0",
    "opentelemetry-instrumentation-alephalpha==0.53.0",
    "opentelemetry-instrumentation-anthropic==0.53.0",
    "opentelemetry-instrumentation-bedrock==0.53.0",
    "opentelemetry-instrumentation-chromadb==0.53.0",
    "opentelemetry-instrumentation-cohere==0.53.0",
    "opentelemetry-instrumentation-crewai==0.53.0",
    "opentelemetry-instrumentation-google-generativeai==0.53.0",
    "opentelemetry-instrumentation-groq==0.53.0",
    "opentelemetry-instrumentation-haystack==0.53.0",
    "opentelemetry-instrumentation-lancedb==0.53.0",
    "opentelemetry-instrumentation-langchain==0.53.0",
    "opentelemetry-instrumentation-llamaindex==0.53.0",
    "opentelemetry-instrumentation-logging==0.61b0",
    "opentelemetry-instrumentation-marqo==0.53.0",
    "opentelemetry-instrumentation-mcp==0.53.0",
    "opentelemetry-instrumentation-milvus==0.53.0",
    "opentelemetry-instrumentation-mistralai==0.53.0",
    "opentelemetry-instrumentation-ollama==0.53.0",
    "opentelemetry-instrumentation-openai==0.53.0",
    "opentelemetry-instrumentation-openai-agents==0.53.0",
    "opentelemetry-instrumentation-pinecone==0.53.0",
    "opentelemetry-instrumentation-qdrant==0.53.0",
    "opentelemetry-instrumentation-redis==0.61b0",
    "opentelemetry-instrumentation-replicate==0.53.0",
    "opentelemetry-instrumentation-requests==0.61b0",
    "opentelemetry-instrumentation-sagemaker==0.53.0",
    "opentelemetry-instrumentation-sqlalchemy==0.61b0",
    "opentelemetry-instrumentation-threading==0.61b0",
    "opentelemetry-instrumentation-together==0.53.0",
    "opentelemetry-instrumentation-transformers==0.53.0",
    "opentelemetry-instrumentation-urllib3==0.61b0",
    "opentelemetry-instrumentation-vertexai==0.53.0",
    "opentelemetry-instrumentation-voyageai==0.53.0",
    "opentelemetry-instrumentation-watsonx==0.53.0",
    "opentelemetry-instrumentation-weaviate==0.53.0",
    "opentelemetry-instrumentation-writer==0.53.0",
    "opentelemetry-proto==1.40.0",
    "opentelemetry-sdk==1.40.0",
    "opentelemetry-semantic-conventions==0.61b0",
    "opentelemetry-semantic-conventions-ai==0.4.15",
    "opentelemetry-util-http==0.61b0",
    "packaging==26.0",
    "pillow==12.1.1",
    "polars==1.38.1",
    "polars-runtime-32==1.38.1",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyjwt==2.11.0",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-engineio==4.13.1",
    "python-multipart==0.0.22",
    "python-socketio==5.16.1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "requests==2.32.5",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "scipy==1.17.1",
    "setuptools==82.0.0",
    "shellingham==1.5.4",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sounddevice==0.5.5",
    "sse-starlette==3.3.2",
    "starlette==0.52.1",
    "sympy==1.14.0",
    "syncer==2.0.3",
    "tenacity==9.1.4",
    "tokenizers==0.22.2",
    "tomli==2.4.0",
    "torch==2.10.0",
    "torchvision==0.25.0",
    "tqdm==4.67.3",
    "traceloop-sdk==0.53.0",
    "triton==3.6.0",
    "typer==0.24.1",
    "types-requests==2.32.4.20260107",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "typing-extensions==4.15.0",
    "ultralytics==8.4.21",
    "ultralytics-thop==2.0.18",
    "urllib3==2.6.3",
    "uvicorn==0.41.0",
    "watchfiles==1.1.1",
    "wrapt==1.17.3",
    "wsproto==1.3.2",
    "yarl==1.23.0",
    "zipp==3.23.0",
]
